Improve preprocessing
